Made significant progress with the custom-comment-tag system.
Have cleaned it up into a few single functions - 40 lines to be exact - but have one issue with how font-lock-keywords plays with the output of the (regexp-opt ...) function.

The current way of turning this on uses a hardcoded copy of the regexp-opt output, except all of the single back-slashes in the original output are replaced by double back-slashes in the current ouput.